+++ RELEASE_NOTES.md
@@ -1,4 +1,12 @@
 # terraform-aviatrix-copilot-init - release notes
 
+## v1.0.1
+- Use data source for login (get new CID on every run)
+- Cleanup requests for readability
+- Correct dependencies
+- Make copilot pwd sensitive
+- Fix readme typo's
+- Add Netflow and Syslog support
+
 ## v1.0.0
 - Initial release